21:53 speech to text changes
